Refactored test_chemistry into two files, one for each function test_chemistry had tests for
