+++ tests/chemistry/test_morgan_fingerprint.py
@@ -1,21 +1,5 @@
-import os
-from pathlib import Path
-
-import pandas as pd
 import pytest
-
-import janitor.chemistry
-
-
-@pytest.mark.parametrize("progressbar", [None, "terminal"])
-@pytest.mark.chemistry
-def test_smiles2mol(chemdf, progressbar):
-    from rdkit import Chem
-
-    chemdf = chemdf.smiles2mol("smiles", "mol", progressbar)
-    assert "mol" in chemdf.columns
-    for elem in chemdf["mol"]:
-        assert isinstance(elem, Chem.rdchem.Mol)
+import janitor.chemistry # noqa: disable=unused-import
 
 
 @pytest.mark.chemistry
